permissions refactor

+++ Modules/Permission/Http/Controllers/Back/RolesController.php
@@ -3,8 +3,8 @@
 namespace Modules\Permission\Http\Controllers\Back;
 
 use Illuminate\Http\Request;
-use Illuminate\Routing\Controller;
 use Spatie\Permission\Models\Role;
+use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Redirect;
 use Modules\Permission\Http\Requests\Back\RoleRequest;
 
@@ -13,12 +13,12 @@ class RolesController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
-
-        $this->middleware('role:admin');
     }
 
     public function index()
     {
+        $this->authorize(request()->route()->getName());
+
         $roles = Role::all();
 
         return view('permission::back.roles.index')->with(compact('roles'));
@@ -26,6 +26,8 @@ class RolesController extends Controller
 
     public function show($role)
     {
+        $this->authorize(request()->route()->getName());
+
         $role = Role::findOrFail($role);
 
         return view('permission::back.roles.show')->with(compact('role'));
@@ -33,11 +35,15 @@ class RolesController extends Controller
 
     public function create()
     {
+        $this->authorize(request()->route()->getName());
+
         return view('permission::back.roles.create', ['role' => new Role()]);
     }
 
     public function store(RoleRequest $request)
     {
+        $this->authorize(request()->route()->getName());
+
         $role = Role::create($request->all());
 
         $role->save();
@@ -49,6 +55,8 @@ class RolesController extends Controller
 
     public function edit($role)
     {
+        $this->authorize(request()->route()->getName());
+
         $role = Role::findOrFail($role);
 
         return view('permission::back.roles.edit')->with(compact('role'));
@@ -56,6 +64,8 @@ class RolesController extends Controller
 
     public function update($role, RoleRequest $request)
     {
+        $this->authorize(request()->route()->getName());
+
         $role = Role::findOrFail($role);
 
         $role->update($request->all());
@@ -67,6 +77,8 @@ class RolesController extends Controller
 
     public function delete($role)
     {
+        $this->authorize(request()->route()->getName());
+
         $role = Role::findOrFail($role);
 
         $role->delete();
